conf: Change virDomainCputune member 'shares' to unsigned long long
cgroup functions set and get the longer type so use it everywhere

+++ src/conf/domain_conf.h
@@ -2011,7 +2011,7 @@ typedef struct _virDomainCputune virDomainCputune;
 typedef virDomainCputune *virDomainCputunePtr;
 
 struct _virDomainCputune {
-    unsigned long shares;
+    unsigned long long shares;
     bool sharesSpecified;
     unsigned long long period;
     long long quota;
